Move date parsing to new file
